// sanity.types.ts
export type PRICING_PLANS_QUERY_RESULT = Array<{
  _id: string;
  priceHeading: string | null;
  chipText: string | null;
  price: string | null;
  features: Array<string> | null;
  ctaLabel: string | null;
  isPopular: boolean | null;
}>;